Add asteroid name reference
This dataset can be used as a reference for the various names that can refer to an asteroid. That will be necessary for us to refer to the same asteroid from different sources with a unique ID. The dataset includes all asteroids from NASA's Small Body Database, which includes over a million asteroids.